Wrote README and refactored project
