+++ README.md
@@ -1,2 +1,19 @@
-# wishlist-api
-API that handles wishlist requests.
+# Wishlist tech stack:
+- SpringBoot
+- MongoDB
+- Redis
+- Kafka
+- Docker
+- Mongock
+- RestAssured
+- Testcontainers
+- Mapstruct
+- Maven
+
+## Local Environment
+- Go to the 'local' directory: `cd local`
+- Execute the start script, to start the app: `./start.sh`
+- To stop the app, simply execute: `./stop.sh`
+
+##  Swagger
+Simply go to `localhost:8080/swagger-ui.html` in order to check out the API contracts
