+++ src/main/java/com/dev/wishlist/rest/controllers/WishlistController.java
@@ -1,8 +1,8 @@
-package com.dev.wishlist.controllers;
+package com.dev.wishlist.rest.controllers;
 
-import com.dev.wishlist.dtos.WishlistResponse;
+import com.dev.wishlist.rest.dtos.WishlistResponse;
 import com.dev.wishlist.models.Product;
-import com.dev.wishlist.openapi.WishlistOpenAPI;
+import com.dev.wishlist.rest.openapi.WishlistOpenAPI;
 import com.dev.wishlist.services.WishlistService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
