+++ src/main/java/com/dev/wishlist/rest/interceptors/RequestInterceptor.java
@@ -1,4 +1,4 @@
-package com.dev.wishlist.interceptors;
+package com.dev.wishlist.rest.interceptors;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
